Add functions for recording convoy routes
The function CC_fnc_convoyRecordRoute works imilarly to unitCapture in
that it sets up a one time only process to record a vehicle trip which
is started and stopped by context menu actions. The CC_fnc_addActions
function simplifies code that adds multiple user actions to an object.
The CC_fnc_arrayFormat function makes it easy to pretty print large
arrays into an easy to work with text format.